update fils afjal

+++ CRICKET_BOOKING_12425/Controllers/API/HeaderNevicationController.cs
@@ -24,25 +24,30 @@ namespace CRICKET_BOOKING_12425.Controllers.API
         {
             try
             {
-                byte[] imageBytes = Convert.FromBase64String(HeaderImg.Imgs);
-
-                string logoDirectory = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "IMG");
-
-                if (!Directory.Exists(logoDirectory))
+                var existingImg = await _dbContext.HeaderImgs.ToListAsync();
+                if (existingImg != null)
                 {
-                    Directory.CreateDirectory(logoDirectory);
+                    return Ok(new { Status = "Exists", Result = "Only one image is allowed." });
                 }
 
-                string fileName = $"Img{DateTime.Now:yyyyMMddHHmmss}.png";
-                string filePath = Path.Combine(logoDirectory, fileName);
+                byte[] imageBytes = Convert.FromBase64String(HeaderImg.Imgs);
 
+                string imageDirectory = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "IMG");
+                if (!Directory.Exists(imageDirectory))
+                {
+                    Directory.CreateDirectory(imageDirectory);
+                }
+
+                string fileName = $"Img_{DateTime.Now:yyyyMMddHHmmssfff}.png";
+                string filePath = Path.Combine(imageDirectory, fileName);
                 await System.IO.File.WriteAllBytesAsync(filePath, imageBytes);
 
                 HeaderImg.Imgs = fileName;
-
                 _dbContext.HeaderImgs.Add(HeaderImg);
                 await _dbContext.SaveChangesAsync();
-                return Ok(new { Status = "Ok", Result = "Save Successfully" });
+
+                return Ok(new { Status = "Ok", Result = "Saved Successfully" });
+
 
             }
             catch (Exception ex)
@@ -52,13 +57,13 @@ namespace CRICKET_BOOKING_12425.Controllers.API
         }
 
         [HttpGet]
-        [Route("ImgDisplay")]
+        [Route("ImgDisplay/{MaineAdminId?}")]
 
-        public async Task<IActionResult> ImgDisplay()
+        public async Task<IActionResult> ImgDisplay(int? MaineAdminId)
         {
             try
             {
-                var Data = await _dbContext.HeaderImgs.ToListAsync();
+                var Data = await _dbContext.HeaderImgs.Where(o=>o.MaineAdminId == MaineAdminId).FirstOrDefaultAsync();
 
                 if(Data != null)
                 {
